renaming folder for readme.md
