+++ src/react-boilerplate/components/terminal/index.jsx
