+++ src/react-boilerplate/components/terminal/styles.scss
